Proxzar/requirements.txt
absl-py 
aiohttp 
aiosignal 
appdirs==1.4.4
asttokens==2.4.1
astunparse 
async-timeout 
attrs 
blinker 
Brotli 
cached-property 
cachetools 
certifi 
cffi 
charset-normalizer 
click 
colorama 
comm==0.2.1
contourpy==1.2.0
cryptography 
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
filelock 
flatbuffers 
fonttools==4.47.2
frozenlist 
fsspec 
gast 
gitdb==4.0.11
GitPython==3.1.41
google-auth 
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio 
h5py 
idna 
importlib-metadata 
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2 
joblib 
jupyter_client==8.6.0
jupyter_core==5.7.1
keras 
Keras-Preprocessing 
kiwisolver==1.4.5
Markdown 
MarkupSafe 
matplotlib==3.8.2
matplotlib-inline==0.1.6
mkl-fft 
mkl-random 
mkl-service==2.4.0
mpmath 
multidict 
nest-asyncio==1.5.9
networkx 
numpy 
oauthlib 
opencv-python-headless==4.8.1.78
opt-einsum 
packaging 
parso==0.8.3
patchify==0.2.3
Pillow 
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf 
psutil==5.9.7
pure-eval==0.2.2
pyasn1 
pyasn1-modules 
pycparser 
Pygments==2.17.2
PyJWT 
pyOpenSSL 
pyparsing==3.1.1
PySocks 
python-dateutil==2.8.2
pyu2f 
pywin32==306
PyYAML 
pyzmq==25.1.2
requests 
requests-oauthlib 
rsa 
scikit-learn 
scipy==1.11.4
sentry-sdk==1.39.2
setproctitle==1.3.3
six 
smmap==5.0.1
stack-data==0.6.3
sympy 
tensorboard 
tensorboard-data-server 
tensorboard-plugin-wit 
tensorflow==2.10.0
tensorflow-estimator 
termcolor 
threadpoolctl 
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
traitlets==5.14.1
typing_extensions 
urllib3 
wandb==0.16.2
wcwidth==0.2.13
Werkzeug 
win-inet-pton 
wrapt 
yarl 
zipp 
